move s.lamp tests to new file #103
